requirements.txt
django==4.2.2
pillow
psycopg2
djangorestframework
python-dotenv